requirements.txt
aiofiles==23.2.1
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
apt-clone==0.2.1
apturl==0.5.2
asgiref==3.8.1
async-timeout==4.0.1
attrs==23.2.0
beautifulsoup4==4.10.0
blinker==1.9.0
Brlapi==0.8.3
Brotli==1.0.9
certifi==2020.6.20
chardet==4.0.0
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
contourpy==1.2.1
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.9
distro==1.7.0
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.15.2
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.1
eyeD3==0.8.10
fastapi==0.111.0
fastapi-cli==0.0.4
ffmpy==0.3.2
filelock==3.18.0
Flask==3.1.0
fonttools==4.53.1
fsspec==2024.6.1
geographiclib==2.0
geopy==2.4.1
gradio==4.36.0
gradio_client==1.0.1
greenlet==3.1.1
gym==0.26.2
gym-notices==0.0.8
gyp==0.1
h11==0.14.0
httpcore==1.0.5
httplib2==0.20.2
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.3
ifaddr==0.1.7
IMDbPY==2021.4.18
importlib-metadata==4.6.4
importlib_resources==6.4.0
iniconfig==2.0.0
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
keyring==23.5.0
kiwisolver==1.4.5
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.1
mdurl==0.1.2
more-itertools==8.10.0
mutagen==1.45.1
nemo-emblems==6.0.1
netaddr==0.8.0
netifaces==0.11.0
numpy==2.0.0
oauthlib==3.2.0
onboard==1.4.1
openai==1.19.0
orjson==3.10.6
packaging==21.3
PAM==0.4.2
pandas==2.2.2
pexpect==4.8.0
Pillow==9.0.1
platformdirs==4.3.6
pluggy==1.5.0
protobuf==3.12.4
psutil==5.9.0
ptyprocess==0.7.0
pycairo==1.20.1
pycryptodomex==3.11.0
pycups==2.0.1
pycurl==7.44.1
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
pyelftools==0.27
pygame==2.6.1
Pygments==2.18.0
PyGObject==3.42.1
PyICU==2.8.1
pyinotify==0.9.6
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyparted==3.11.7
pyRFC3339==1.1
pytest==8.2.0
python-apt==2.4.0+ubuntu2
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
python-gnupg==0.4.8
python-magic==0.4.24
python-multipart==0.0.9
python-xlib==0.29
pytube==15.0.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
qrcode==7.3.1
referencing==0.35.1
regex==2024.5.15
reportlab==3.6.8
requests==2.25.1
requests-file==1.5.1
rich==13.7.1
rpds-py==0.18.1
ruff==0.5.1
safetensors==0.4.3
SecretStorage==3.3.1
semantic-version==2.10.0
setproctitle==1.2.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.3.1
SQLAlchemy==2.0.39
sqlparse==0.5.3
starlette==0.37.2
systemd-python==234
tinycss2==1.1.1
tldextract==3.1.2
tokenizers==0.19.1
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.1
tqdm==4.66.4
transformers==4.40.0
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
ubuntu-drivers-common==0.0.0
ufw==0.36.1
ujson==5.10.0
Unidecode==1.3.3
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
virtualenv==20.29.3
wadllib==1.3.6
watchfiles==0.22.0
webencodings==0.5.1
webrtcvad==2.0.10
websockets==11.0.3
Werkzeug==3.1.3
xdg==5
xkit==0.0.0
xlrd==1.2.0
yt-dlp==2022.4.8
zipp==1.0.0
